\name{caper-benchmarks}
\alias{caper-benchmarks}
\alias{benchBayesTraitsOutputs}
\alias{benchBrunchOutputs}
\alias{benchCrunchOutputs}
\alias{benchFuscoOutputs}
\alias{benchMacroCAICOutputs}
\alias{benchMesaOutputs}
\alias{benchTestInputs}
\alias{BayesTraitsMods}
\alias{benchData}
\alias{benchTreeDicho}
\alias{benchTreePoly}
\alias{CAIC.BrDi1057}
\alias{CAIC.BrDi1157}
\alias{CAIC.BrDi813}
\alias{CAIC.BrDi913}
\alias{CAIC.BrPl1057}
\alias{CAIC.BrPl1157}
\alias{CAIC.BrPl813}
\alias{CAIC.BrPl913}
\alias{CAIC.CrDi213}
\alias{CAIC.CrDi657}
\alias{CAIC.CrPl213}
\alias{CAIC.CrPl413}
\alias{CAIC.CrPl657}
\alias{FuscoDiSpp}
\alias{FuscoDiTax}
\alias{FuscoPolySpp}
\alias{FuscoPolyTax}
\alias{MacroCAIC.DiSpp23}
\alias{MacroCAIC.DiSpp67}
\alias{MacroCAIC.DiTax23}
\alias{MacroCAIC.DiTax67}
\alias{MacroCAIC.PolySpp23}
\alias{MacroCAIC.PolySpp67}
\alias{MacroCAIC.PolyTax23}
\alias{MacroCAIC.PolyTax67}
\alias{MeSA.I}
\alias{testData}
\alias{testTree}
\alias{nul}
\alias{fix}
\alias{kld}
\alias{Kld}
\alias{kLd}
\alias{klD}
\alias{KLd}
\alias{kLD}
\alias{KlD}
\alias{KLD}

\docType{data}
\title{Datasets used for benchmarking caper}
\description{
These data files and model objects contain input datasets for benchmarking the caper package against other implementations and the results returned by those other implementations. For more details, see the benchmark vignette for caper.
}
\seealso{vignette()}
\keyword{datasets}
