\name{fuscoData}
\alias{fuscoBirdData}
\alias{fuscoBirdTree}

\title{Example dataset for Fusco imbalance calculations}
\description{
This dataset contains the phylogeny of bird families and species richness originally included with the FUSCO imbalance calculation programs.
}
\usage{
data(fuscoData)
}
\format{
The dataset provides a phylogeny (\code{fuscoBirdTree}) and a data frame (\code{fuscoBirdData}). The phylogeny is a 137 tip tree containing one polytomy and the dataframe provides tip labels and species richness for each bird family. This dataset was provided with the original DOS implementation of the test and the families were unlabelled in this original file.
}
\seealso{fusco.test}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}


