% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caic.methods.R
\name{summary.caic}
\alias{summary.caic}
\alias{print.caic}
\title{Summarize a crunch, brunch or macrocaic analysis}
\usage{
\method{summary}{caic}(object, ...)
}
\arguments{
\item{object}{An object of class 'caic'.}

\item{\dots}{Arguments to be passed to 'summary.lm'.}

\item{x}{An object of class 'caic'.}
}
\value{
The summary method returns an object of class 'summary.lm'.
}
\description{
The summary method simply returns the linear model summary from the 'caic'
object. The print method prints some basic information about the analysis
followed by the model summary.
}
\examples{

data(shorebird)
shorebird <- comparative.data(shorebird.tree, shorebird.data, Species)
crunchMod <- crunch(Egg.Mass ~ F.Mass + M.Mass, data=shorebird)
print(crunchMod)
summary(crunchMod)

}
\seealso{
\code{\link{crunch}},\code{\link{brunch}}, \code{link{macrocaic}}
}
\author{
David Orme
}
\keyword{methods}
