\name{syrphidae}
\alias{syrphidae}
\alias{syrphidaeTree}
\alias{syrphidaeRich}


\title{The syrphidae dataset of Katzourakis et al. 2001 }
\description{
A genus level phylogeny of the Syrphidae (sawflies) along with data on the species richness of each genus.
}
\usage{
data(syrphidae)
}
\format{
A 'phylo' object (\code{syrphidaeTree}) containing a phylogeny of 204 genera of sawflies and a data frame  (\code{syrphidaeRich}) of the species richness of each genus. 
}

\references{
Katzourakis, A., Purvis, A., Azmeh, S., Rotheray, G., and Gilbert, F. (2001). Macroevolution of hoverflies (Diptera : Syrphidae): the effect of using higher-level taxa in studies of biodiversity, and correlates of species richness. Journal of Evolutionary Biology, 14:219-227.
}

\seealso{fusco.test}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}

