\name{caper-package}
\alias{caper-package}
\alias{caper}

\docType{package}
\title{
Comparative analysis of phylogenetics and evolution in R (caper)
}
\description{The \pkg{caper} package provides a set of functions to conduct comparative analyses using both independent contrasts and phylogenetic generalised least squares methods. It also provides functions to combine phylogeny objects with data frames into simple comparative datasets and some utility functions for manipulating phylogenies.

In addition to linear models correcting for phylogeny, the package also provides functions to test tree imbalance, calculate various measures of phylogenetic diversity and simulate phylogenies and traits.
}
\details{
\tabular{ll}{
Package: \tab caper\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2011-05-06\cr
License: \tab GPL\cr
}
A package to carry out phylogenetic comparative analysis.
}
\author{
David Orme
Maintainer: David Orme <d.orme@imperial.ac.uk>
}
\keyword{ package }
\seealso{
\code{\link[ape:ape-package]{ape}}
}
