\name{syrphidae}
\alias{syrphidae}
\alias{syrphidaeTree}
\alias{syrphidaeRich}


\title{The syrphidae dataset of Purvis et al. 2002 }
\description{
A genus level phylogeny of the Syrphidae (sawflies) along with data on the species richness of each genus.
}
\usage{
data(syrphidae)
}
\format{
A 'phylo' object (\code{syrphidaeTree}) containing a phylogeny of 204 genera of sawflies and a data frame  (\code{syrphidaeRich}) of the species richness of each genus. 
}

\references{
Purvis A.,  Katzourakis A. & Agapow, P-M (2002) Evaluating Phylogenetic Tree Shape: Two Modifications to Fusco & Cronk's Method. J. Theor. Biol. 214, 93-103.
}

\seealso{fusco.test}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{datasets}

